Changed no floor tiles to nomove
Tiles floating in the air on plane > 0 are actually walkable, but you need to noclip to get there. In certain edge cases, usually involving transports, it is possible for the pathfinder to traverse such no floor tiles. Therefore, they are now turned into nomove tiles if the minimap tile colour is black